BPR-50:
Updated the Intake table in the database to also include the progress made on calorie and water intake for the day, as a percentage of the goal set.

Also updated the UI to include a progress bar and made the input fields not move around as much when adding calories or water to intake.